chore: add flow model for postgres
