+++ internal/store/postgres/session.go
@@ -15,15 +15,6 @@ type Session struct {
 	CreatedAt       time.Time `db:"created_at"`
 }
 
-type Flow struct {
-	ID        uuid.UUID `db:"id"`
-	Method    string    `db:"method"`
-	StartURL  string    `db:"start_url"`
-	FinishURL string    `db:"finish_url"`
-	Nonce     string    `db:"nonce"`
-	CreatedAt time.Time `db:"created_at"`
-}
-
 func (s *Session) transformToSession() *authenticate.Session {
 	return &authenticate.Session{
 		ID:              s.ID,
@@ -31,16 +22,5 @@ func (s *Session) transformToSession() *authenticate.Session {
 		AuthenticatedAt: s.AuthenticatedAt,
 		ExpiresAt:       s.ExpiresAt,
 		CreatedAt:       s.CreatedAt,
-	}
-}
-
-func (f *Flow) transformToFlow() *authenticate.Flow {
-	return &authenticate.Flow{
-		ID:        f.ID,
-		Method:    f.Method,
-		StartURL:  f.StartURL,
-		FinishURL: f.FinishURL,
-		Nonce:     f.Nonce,
-		CreatedAt: f.CreatedAt,
 	}
 }
